requirements.txt
sqlite
pycryptodomex
pywin32